+++ src/solver/model_getters.jl
@@ -214,6 +214,8 @@ function get_radiation_mode(parsed_args, ::Type{FT}) where {FT}
     @assert idealized_insolation in (true, false)
     idealized_clouds = parsed_args["idealized_clouds"]
     @assert idealized_clouds in (true, false)
+    add_isothermal_boundary_layer = parsed_args["add_isothermal_boundary_layer"]
+    @assert add_isothermal_boundary_layer in (true, false)
     radiation_name = parsed_args["rad"]
     @assert radiation_name in (
         nothing,
@@ -230,24 +232,28 @@ function get_radiation_mode(parsed_args, ::Type{FT}) where {FT}
             idealized_h2o,
             idealized_insolation,
             idealized_clouds,
+            add_isothermal_boundary_layer,
         )
     elseif radiation_name == "gray"
         RRTMGPI.GrayRadiation(
             idealized_h2o,
             idealized_insolation,
             idealized_clouds,
+            add_isothermal_boundary_layer,
         )
     elseif radiation_name == "allsky"
         RRTMGPI.AllSkyRadiation(
             idealized_h2o,
             idealized_insolation,
             idealized_clouds,
+            add_isothermal_boundary_layer,
         )
     elseif radiation_name == "allskywithclear"
         RRTMGPI.AllSkyRadiationWithClearSkyDiagnostics(
             idealized_h2o,
             idealized_insolation,
             idealized_clouds,
+            add_isothermal_boundary_layer,
         )
     elseif radiation_name == "DYCOMS_RF01"
         RadiationDYCOMS_RF01{FT}()